02_meme_gen_starter/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
pycodestyle==2.5.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
validators==0.14.2
Werkzeug==1.0.0
